jwt service add

+++ service/map.service.js
@@ -1,11 +1,23 @@
 const map = require("../repository/map.repository.js");
+const data = require("../utils/getData_utils");
 
-exports.applyMap = (req, res) => {
+exports.applyMap = async (req, res) => {
   if (!req.body) {
     res.status(400).send({
       message: "req.body can not be empty!",
     });
   } else {
+    if (req.body.map_maker) {
+      const temp = await data.findUserById(req.body.map_maker);
+      if(temp.ok){
+          console.log(req.body.map_maker);
+      } else {
+        res.status(400).send({
+          message: "applyMap err",
+        });
+        return;
+      }
+    }
     map.createMap(req.body, (err, result) => {
       if (err) {
         res.status(400).send({
@@ -35,7 +47,7 @@ exports.getAllMapList = (req, res) => {
 exports.getMapListByTag = (req, res) => {
   if (!req.params) {
     res.status(400).send({
-      message: "req.body can not be empty!",
+      message: "req.params can not be empty!",
     });
   } else {
     map.findAllMapsByTag(req.params.map_tag, (err, result) => {
@@ -70,28 +82,32 @@ exports.updateMapInfo = (req, res) => {
   }
 };
 
-exports.deleteMap = (req, res) => {
+exports.deleteMap = async (req, res) => {
   if (!req.params) {
     res.status(400).send({
       message: "req.params can not be empty!",
     });
-  }
-//   else if (!req.user) {
-//     res.status(401).send({
-//       message: "Unauthorized",
-//     });
-//   }
-  else {
-    map.deleteMap(req.params.map_id, (err, results) => {
-      if (err) {
-        res.status(400).send({
-          message: "deleteMap err",
-        });
-      } else {
-        res.send({
-          message: `map(${req.params.map_id}) was successfully deleted`,
-        });
-      }
-    });
+  } else {
+    const forDelete = await data.findMapById(req.params.map_id)
+    if (forDelete.ok === false) {
+      res.status(400).send({ message: `There is no map(${req.params.map_id})` });
+    }
+    else if (req.user !== forDelete.map.map_maker) {
+      res.status(401).send({
+          message: "Unauthorized. Can only delete my map."
+      });
+    } else {
+      map.deleteMap(req.params.map_id, (err, results) => {
+        if (err) {
+          res.status(400).send({
+            message: "deleteMap err",
+          });
+        } else {
+          res.send({
+            message: `map(${req.params.map_id}) was successfully deleted`,
+          });
+        }
+      });
+    }
   }
 };
